Completed JavaDoc, updated README and removed files.
The JavaDoc is now completed, so a pull-request can be done for merging this branch into the "development" one.
The enums for the database values have been removed as they are unefficient with the Android resources (see: https://android.jlelse.eu/android-performance-avoid-using-enum-on-android-326be0794dc3) so now they are just static strings with useful contents.
Travis-CI has been removed and replaced by GitLab-CI (this is the first on-working commit to the new platform), so now the Continuous Integration is going to be done on the GitLab-Runner platform. For that reason, ".travis.yml" has been replaced by ".gitlab-ci.yml".
Finally, README has been updated with the new CI values and some libraries in Gradle have been upgraded.